# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Syndikat Disc Golf API
  version: 1.0.0
  description: |
    Website-facing API for syndikat.golf. This contract is maintained manually alongside Express routes.
    Third-party scraper data can change independently; clients must tolerate omitted optional fields.
  license:
    name: ISC
    identifier: ISC
  contact:
    name: Syndikat Disc Golf
    url: https://syndikat.golf
servers:
  - url: https://api.syndikat.golf
    description: Production
  - url: http://{host}:8080
    description: Local development
    variables:
      host:
        default: localhost
# All endpoints are public unless an operation overrides this with a security requirement.
security: []
...
paths:
...
  /ratings:
    get:
      tags: [Community]
      summary: Return paginated player ratings
      description: Returns the latest player ratings collected from the source site. Results can be filtered and paginated with query parameters.
      operationId: getPlayerRatings
      parameters:
        - name: club
          in: query
          description: Case-insensitive club substring filter.
          schema: { type: string }
        - name: division
          in: query
          description: Exact division filter.
          schema: { type: string }
        - name: search
          in: query
          description: Case-insensitive search across player name and club.
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Ratings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedRatings' }
        '500':
          $ref: '#/components/responses/ServerError'
  /ratings/{club}:
    get:
      tags: [Community]
      summary: Return player ratings filtered by club
      description: Case-insensitive substring match on club names, treating Disc Golf, Discgolf, Disc-Golf, and DG as equivalent. Whitespace is normalized. Empty matches return an empty array. Rankings and division counts remain those of the full ratings list.
      operationId: getPlayerRatingsByClub
      parameters:
        - name: club
          in: path
          required: true
          description: Club name or part of it. URL-encode spaces.
          schema:
            type: string
          example: disc golf syndikat
      responses:
        '200':
          description: Matching ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rating'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /tournaments/route:
    post:
      tags: [Tournaments]
      summary: Plan a driving route to a tournament
      description: The origin is geocoded in Germany. A Deutsche Bahn URL is included only when station lookup is configured.
      operationId: planTournamentRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '200':
          description: Driving route and resolved locations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Invalid request, including malformed JSON, or an unresolvable origin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteError'
        '404':
          description: No driving route found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteError'
        '502':
          description: Routing provider error or quota exhaustion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteError'
  /products/feed:
    get:
      tags: [Products]
      summary: Return recent products from supported Shopify feeds
      description: Returns recent products from every supported Shopify feed.
      operationId: getProductFeed
      responses:
        '200':
          description: Products ordered newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          $ref: '#/components/responses/ServerError'
  /out/{shopId}/{productPath}:
    get:
      tags: [Products]
      summary: Redirect to a validated merchant product URL
      description: Resolves the readable product path against the configured shop origin, adds the Syndikat referral parameter, and returns a temporary redirect. The productPath wildcard may contain multiple slash-separated segments. Valid redirects emit a best-effort product_click event unless the syndikat_analytics_disabled=1 cookie is present.
      operationId: redirectToProduct
      parameters:
        - name: shopId
          in: path
          required: true
          schema: { type: string }
        - name: productPath
          in: path
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect to the configured merchant host
          headers:
            Location: { schema: { type: string, format: uri } }
        '400':
          description: Product path is invalid for the configured shop
        '404':
          description: Invalid shop
  /products/search-stream/{query}:
    get:
      tags: [Products]
      summary: Search all supported stores as a server-sent event stream
      description: |
        The response uses `text/event-stream`. A `start` event is followed by zero or more
        `store`/`store-error` events and one `end` event. The server sends comment heartbeats.
        Closing the connection cancels outstanding store requests.
      operationId: streamProductSearch
      parameters:
        - $ref: '#/components/parameters/ProductQuery'
      responses:
        '200':
          description: Search event stream
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                start:
                  value: "event: start\ndata: {\"query\":\"Buzzz\",\"stores\":[\"discgolfstore\"]}\n\n"
                store:
                  value: "event: store\ndata: {\"store\":\"discgolfstore\",\"products\":[],\"count\":0}\n\n"
      x-sse-events:
        - event: start
          schema:
            $ref: '#/components/schemas/ProductSearchStart'
        - event: store
          schema:
            $ref: '#/components/schemas/ProductSearchStore'
        - event: store-error
          schema:
            $ref: '#/components/schemas/ProductSearchStoreError'
        - event: end
          schema:
            $ref: '#/components/schemas/ProductSearchEnd'
        - event: error
          schema:
            $ref: '#/components/schemas/Message'
  /products/{type}/{query}:
    get:
      tags: [Products]
      summary: Search a single supported store
      description: Searches one supported store for products matching the query.
      operationId: searchStoreProducts
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
          description: Store identifier. Unsupported identifiers return a text diagnostic.
        - $ref: '#/components/parameters/ProductQuery'
      responses:
        '200':
          description: Products, no result, or an unsupported-store diagnostic
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  - type: 'null'
                  - type: string
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    RemovalToken:
      type: http
      scheme: bearer
      bearerFormat: opaque removal token
  parameters:
    ProductQuery:
      name: query
      in: path
      required: true
      schema:
        type: string
      description: URL-encoded search terms.
  responses:
...
    ServerError:
      description: Server or upstream error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
...
  schemas:
...
    Message:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    Rating:
      type: object
      required: [firstName, lastName, rating, ratingChange, divisionCount, gtNumber, division, lastRound, roundCount, rank, divisionRank, link, club]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        rating: { type: number }
        ratingChange: { type: [integer, 'null'], enum: [1, -1, null] }
        divisionCount: { type: integer }
        gtNumber: { type: integer }
        division: { type: string }
        lastRound: { type: string, format: date-time }
        roundCount: { type: integer }
        rank: { type: integer }
        divisionRank: { type: integer }
        link: { type: string }
        club: { type: string }
    PaginatedRatings:
      type: object
      required: [items, total, page, pageSize, divisions]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Rating' }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        divisions:
          type: array
          items: { type: string }
...
    RouteRequest:
      type: object
      required: [origin, destination]
      properties:
        origin: { type: string, minLength: 3 }
        date: { type: string, minLength: 1 }
        destination:
          type: object
          required: [lat, lng]
          properties:
            lat: { type: number, minimum: -90, maximum: 90 }
            lng: { type: number, minimum: -180, maximum: 180 }
            address: { type: string, minLength: 1 }
            name: { type: string, minLength: 1 }
            station: { type: string, minLength: 1 }
    RouteResponse:
      type: object
      required: [route, resolvedPoints]
      properties:
        route:
          type: object
          required: [geometry, distanceMeters, durationSeconds]
          properties:
            geometry:
              type: object
              required: [type, coordinates]
              properties:
                type: { type: string }
                coordinates:
                  type: array
                  items:
                    type: array
                    prefixItems: [{ type: number }, { type: number }]
                    minItems: 2
                    maxItems: 2
            distanceMeters: { type: number }
            durationSeconds: { type: number }
        resolvedPoints:
          type: object
          required: [origin, destination]
          properties:
            origin: { $ref: '#/components/schemas/ResolvedPoint' }
            destination: { $ref: '#/components/schemas/ResolvedPoint' }
        train:
          type: object
          required: [url]
          properties:
            url: { type: string, format: uri }
    ResolvedPoint:
      type: object
      required: [text, lat, lng]
      properties:
        text: { type: string }
        lat: { type: number }
        lng: { type: number }
    RouteError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [INVALID_REQUEST, INVALID_ADDRESS, NO_ROUTE_FOUND, UPSTREAM_ERROR, UPSTREAM_QUOTA_EXCEEDED] }
        message: { type: string }
    Product:
      type: object
      required: [title, price, store, url, stockStatus, crawledAt]
      properties:
        title: { type: string }
        price: { type: [number, 'null'] }
        image: { type: [string, 'null'] }
        store: { type: string }
        vendor: { type: string }
        url: { type: [string, 'null'], format: uri }
        flightNumbers:
          type: object
          properties:
            speed: { type: [string, 'null'] }
            glide: { type: [string, 'null'] }
            turn: { type: [string, 'null'] }
            fade: { type: [string, 'null'] }
        stockStatus: { type: string, enum: [available, unavailable, unknown] }
        crawledAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    ProductSearchStart:
      type: object
      required: [query, stores]
      properties:
        query: { type: string }
        stores: { type: array, items: { type: string } }
    ProductSearchStore:
      type: object
      required: [store, products, count]
      properties:
        store: { type: string }
        products: { type: array, items: { $ref: '#/components/schemas/Product' } }
        count: { type: integer, minimum: 0 }
    ProductSearchStoreError:
      type: object
      required: [store, message]
      properties:
        store: { type: string }
        message: { type: string }
    ProductSearchEnd:
      type: object
      required: [query, storesCompleted, totalProducts, failedStores]
      properties:
        query: { type: string }
        storesCompleted: { type: integer, minimum: 0 }
        totalProducts: { type: integer, minimum: 0 }
        failedStores: { type: array, items: { type: string } }
...